+++ _putchar.c
@@ -1,0 +1,11 @@
+#include <unistd.h>
+
+/**
+ * _putchar - write the character c to stdout
+ * @c: The character to print
+ * Return: on success 1.
+ */
+int _putchar(char c)
+{
+	return (write(1, &c, 1));
+}
