+++ environ.c
