+++ realloc.c
@@ -1,11 +1,11 @@
 #include "shell.h"
 
 /**
- *memeset fills with a 
+ * _memset - fills memory with a constant byte
  * @s: the pointer to the memory area
- * @b: the by
+ * @b: the byte to fill *s with
  * @n: the amount of bytes to be filled
- * Return: (s) a 
+ * Return: (s) a pointer to the memory area s
  */
 char *_memset(char *s, char b, unsigned int n)
 {
