+++ shell.h
@@ -51,10 +51,9 @@ typedef struct liststr
 } list_t;
 
 /**
- * struct passinfo - contains pseudo-arguements to pass into a function,
- * allowing uniform prototype for function pointer struct
- * @arg: a string generated from getline containing arguements
- * @argv:an array of strings generated from arg
+ * struct passinfo - Pseudo-arguements function,
+ * @arg: string gen from getline 
+ * @argv:array of strings gen from arg
  * @path: a string path for the current command
  * @argc: the argument count
  * @line_count: the error count
@@ -62,15 +61,15 @@ typedef struct liststr
  * @linecount_flag: if on count this line of input
  * @fname: the program filename
  * @env: linked list local copy of environ
- * @environ: custom modified copy of environ from LL env
+ * @environ: custm LL env
  * @history: the history node
- * @alias: the alias node
- * @env_changed: on if environ was changed
- * @status: the return status of the last exec'd command
- * @cmd_buf: address of pointer to cmd_buf, on if chaining
+ * @alias: alias 
+ * @env_changed: environ changed
+ * @status: the exec'd command
+ * @cmd_buf: pointer to cmd buffer 
  * @cmd_buf_type: CMD_type ||, &&, ;
- * @readfd: the fd from which to read line input
- * @histcount: the history line number count
+ * @readfd: fd input
+ * @histcount:count
  */
 typedef struct passinfo
 {
@@ -89,8 +88,8 @@ typedef struct passinfo
 	int env_changed;
 	int status;
 
-	char **cmd_buf; /* pointer to cmd ; chain buffer, for memory mangement */
-	int cmd_buf_type; /* CMD_type ||, &&, ; */
+	char **cmd_buf; 
+	int cmd_buf_type;
 	int readfd;
 	int histcount;
 } info_t;
@@ -100,9 +99,9 @@ typedef struct passinfo
 		0, 0, 0}
 
 /**
- * struct builtin - contains a builtin string and related function
- * @type: the builtin command flag
- * @func: the function
+ * struct builtin - builtin string 
+ * @type: command flag
+ * @func: funct
  */
 typedef struct builtin
 {
@@ -232,4 +231,4 @@ int replace_alias(info_t *);
 int replace_vars(info_t *);
 int replace_string(char **, char *);
 
-#endif
+#endif
