+++ shell_loop.c
@@ -1,11 +1,11 @@
 #include "shell.h"
 
 /**
- * hsh - main shell file
- * @info:  parameter & return 
- * @av:  argument vector  main()
+ * hsh - main shell loop
+ * @info: the parameter & return info struct
+ * @av: the argument vector from main()
  *
- * Return: 0 on success, 1 on error code
+ * Return: 0 on success, 1 on error, or error code
  */
 int hsh(info_t *info, char **av)
 {
@@ -44,13 +44,13 @@ int hsh(info_t *info, char **av)
 }
 
 /**
- * find_builtin
+ * find_builtin - finds a builtin command
  * @info: the parameter & return info struct
  *
- * Return: -1 if builtin not found,
- * 	0 if builtin comment
- * 	1 if  successful,
- * 	2 if builtin ()
+ *Return: -1  not found,
+ * 0 ,
+ * 1 not successful,
+ * 2 signals exit()
  */
 int find_builtin(info_t *info)
 {
@@ -78,8 +78,8 @@ int find_builtin(info_t *info)
 }
 
 /**
- * find_cmd -to PATH
- * @info: the to struct
+ * find_cmd - finds a command in PATH
+ * @info: the parameter & return info struct
  *
  * Return: void
  */
@@ -120,10 +120,10 @@ void find_cmd(info_t *info)
 }
 
 /**
- * fork_cmd - forks  to run cmd
- * @info: the to
+ * fork_cmd - forks a an exec thread to run cmd
+ * @info: the parameter & return info struct
  *
- * Return: void null 
+ * Return: void
  */
 void fork_cmd(info_t *info)
 {
@@ -158,3 +158,4 @@ void fork_cmd(info_t *info)
 		}
 	}
 }
+
