+++ string.c
@@ -1,10 +1,10 @@
 #include "shell.h"
 
 /**
- * _strlen - returns a string
- * @s:  length to check
+ * _strlen - returns the length of a string
+ * @s: the string whose length to check
  *
- * Return: integer of string
+ * Return: integer length of string
  */
 int _strlen(char *s)
 {
@@ -19,9 +19,9 @@ int _strlen(char *s)
 }
 
 /**
- * _strcmp .
+ * _strcmp - performs lexicogarphic comparison of two strangs.
  * @s1: the first strang
- * @s2:  strang
+ * @s2: the second strang
  *
  * Return: negative if s1 < s2, positive if s1 > s2, zero if s1 == s2
  */
@@ -42,8 +42,8 @@ int _strcmp(char *s1, char *s2)
 
 /**
  * starts_with - checks if needle starts with haystack
- * @haystack: string 
- * @needle: the substring  find
+ * @haystack: string to search
+ * @needle: the substring to find
  *
  * Return: address of next char of haystack or NULL
  */
