+++ vars.c
@@ -1,12 +1,12 @@
 #include "shell.h"
 
 /**
- * is_chain - test 
- * @info: the
- * @buf:  buffer
- * @p: address in buf
+ * is_chain - delimeter
+ * @info: the parameter struct
+ * @buf: buffer
+ * @p: buf
  *
- * Return: 1 if r, 0 otherwise
+ * Return: 1 if chain delimeter, 0 otherwise
  */
 int is_chain(info_t *info, char *buf, size_t *p)
 {
@@ -36,12 +36,12 @@ int is_chain(info_t *info, char *buf, size_t *p)
 }
 
 /**
- * check status
- * @info:  parameter
- * @buf:  char r
- * @p:  buf
- * @i:  buf
- * @len: l buf
+ * check_chain - checks we should continue chaining based on last status
+ * @info: the parameter struct
+ * @buf: the char buffer
+ * @p: address of current position in buf
+ * @i: starting position in buf
+ * @len: length of buf
  *
  * Return: Void
  */
@@ -70,8 +70,8 @@ void check_chain(info_t *info, char *buf, size_t *p, size_t i, size_t len)
 }
 
 /**
- * replace string
- * @info: parameter
+ * replace_alias - replaces an aliases in the tokenized string
+ * @info: the parameter struct
  *
  * Return: 1 if replaced, 0 otherwise
  */
@@ -100,9 +100,9 @@ int replace_alias(info_t *info)
 
 /**
  * replace_vars - replaces vars in the tokenized string
- * @info: parameter
+ * @info: the parameter struct
  *
- * Return: 1 if
+ * Return: 1 if replaced, 0 otherwise
  */
 int replace_vars(info_t *info)
 {
@@ -140,8 +140,8 @@ int replace_vars(info_t *info)
 }
 
 /**
- * replace string
- * @old: address
+ * replace_string - replaces string
+ * @old: address of old string
  * @new: new string
  *
  * Return: 1 if replaced, 0 otherwise
